+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Minor changes and fixes
 
+* Avatar set for anonymous users: new 'none' choice (that will fallback to Converse new colorized avatars).
 * New translation: Albanian.
 * Translation updates: Crotian, Japanese.